Move fetching of profile data to the DashboardProvider so it doesn't have to be duplicated in the DashboardPage and the ShoppingListPage
