+++ src/pages/shoppingListPage/shoppingListPage.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useRef }from 'react'
+import React, { useState, useEffect }from 'react'
 import { Redirect } from 'react-router-dom'
 import { backendBaseUri } from '../../utils/config'
 import colorSchemes, { YELLOW } from '../../utils/colorSchemes'
@@ -24,46 +24,7 @@ const ShoppingListPage = () => {
   const [flashVisible, setFlashVisible] = useState(false)
   const [shouldRedirect, setShouldRedirect] = useState(false)
 
-  const mountedRef = useRef(true)
-
-  const { token, setProfileData, removeSessionCookie } = useDashboardContext()
-
-  const fetchProfileData = () => {
-    const dataUri = `${backendBaseUri[process.env.NODE_ENV]}/users/current`
-
-    if (mountedRef.current === true && (isStorybook() || !!token)) {
-      fetch(dataUri, {
-        headers: {
-          'Content-Type': 'application/json',
-          'Authorization': `Bearer ${token}`
-        }
-      })
-        .then(response => (response.json()))
-        // TODO: https://trello.com/c/JRyN8FSN/25-refactor-error-handling-in-promise-chains
-        .then(data => {
-          if (!data) {
-            console.warn('No data received from server - logging out user just in case')
-            token && removeSessionCookie()
-            setShouldRedirect(true)
-          } else if (data.error) {
-            console.warn('Error fetching user data - logging out user: ', data.error)
-            token && removeSessionCookie()
-            setShouldRedirect(true)
-          } else {
-            setProfileData(data)
-            setShouldRedirect(false)
-          }
-        })
-        .catch(error => {
-          console.error('Error returned while fetching profile data: ', error.message)
-          token && removeSessionCookie()
-          setShouldRedirect(true)
-        })
-    } else {
-      token && removeSessionCookie()
-      setShouldRedirect(true)
-    }
-  }
+  const { token, removeSessionCookie } = useDashboardContext()
 
   const fetchLists = () => {
     const dataUri = `${backendBaseUri[process.env.NODE_ENV]}/shopping_lists`
@@ -173,7 +134,6 @@ const ShoppingListPage = () => {
   }
 
   useEffect(fetchLists, [])
-  useEffect(fetchProfileData, [])
 
   return(
     <DashboardLayout title='Your Shopping Lists'>
